requirements.txt
# OfferCompare Pro - Dependencies
# Core framework and LLM integration
pocketflow>=0.0.1

# Environment management
python-dotenv>=1.0.0

# AI Provider APIs
openai>=1.0.0
google-genai>=1.0.0
anthropic>=0.7.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization and charts
matplotlib>=3.7.0
seaborn>=0.12.0

# Web scraping and requests (for future API integrations)
requests>=2.31.0
beautifulsoup4>=4.12.0

# JSON and data validation
pydantic>=2.0.0
jsonschema>=4.17.0

# Type hints support  
typing-extensions>=4.5.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Backend API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Auth and rate limiting (Supabase). PyJWT installs as 'jwt' (import jwt).
supabase>=2.0.0
PyJWT>=2.8.0